refactor

+++ src/main/java/it/uniroma3/siwFood/service/IngredientService.java
@@ -3,9 +3,15 @@ package it.uniroma3.siwFood.service;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
+import it.uniroma3.siwFood.model.Cook;
+import it.uniroma3.siwFood.model.Credentials;
 import it.uniroma3.siwFood.model.Ingredient;
+import it.uniroma3.siwFood.model.Recipe;
 import it.uniroma3.siwFood.repository.IngredientRepository;
 
 @Service
@@ -13,6 +19,12 @@ public class IngredientService {
 	
 	@Autowired
 	private IngredientRepository ingredientRepository;
+	
+	@Autowired
+	private CredentialsService credentialsService;
+	
+	@Autowired
+	private CookService cookService;
 	
 	public List<Ingredient> findIngredientsByRecipeId(Long idRecipe){
 		
@@ -24,13 +36,39 @@ public class IngredientService {
 		return this.ingredientRepository.findById(idIngredient).get();
 	}
 	
-	public void saveIngredient(Ingredient ingredient) {
+	/*VIENE SALVATO UN NUOVO INGREDIENTE SOLO SE È RICHIESTO DAL CUOCO PROPRIETARIO DELLA RICETTA
+	 * OPPURE DA UN ADMIN*/
+	public void saveIngredient(Ingredient ingredient) throws AccessDeniedException {
 		
-		this.ingredientRepository.save(ingredient);
+		Recipe recipe = ingredient.getRecipe();
+	    UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		Credentials credentials = this.credentialsService.findCredenzialiByUsername(userDetails.getUsername());
+		Cook cook = this.cookService.findCookByCredentials(credentials.getIdCredentials());
+		
+		if(recipe.getCook().equals(cook) || credentials.getRole().equals("ADMIN")) {
+			
+			this.ingredientRepository.save(ingredient);
+		}				
+		else {
+			throw new AccessDeniedException("You do not have permission to add ingredients");
+		}
 	}
 	
-	public void deleteIngredient(Ingredient ingredient) {
+	/*VIENE ELIMINATO UN INGREDIENTE, DA UNA RICETTA, SOLO SE LO RICHIEDE IL CUOCO PROPRIETARIO DELLA
+	 * RICETTA OPPURE UN ADMIN*/
+	public void deleteIngredient(Ingredient ingredient) throws AccessDeniedException {
 		
-		this.ingredientRepository.delete(ingredient);
+		Recipe recipe = ingredient.getRecipe();
+	    UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		Credentials credentials = this.credentialsService.findCredenzialiByUsername(userDetails.getUsername());
+		Cook cook = this.cookService.findCookByCredentials(credentials.getIdCredentials());
+		
+		if(recipe.getCook().equals(cook) || credentials.getRole().equals("ADMIN")) {
+			
+			this.ingredientRepository.delete(ingredient);
+		}				
+		else {
+			throw new AccessDeniedException("You do not have permission to delete ingredients");
+		}
 	}
 }
